+++ lib/src/middleware/require_auth.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-import 'dart:io';
 import 'package:angel_framework/angel_framework.dart';
 
 /// Forces Basic authentication over the requested resource, with the given [realm] name, if no JWT is present.
@@ -10,27 +8,32 @@ RequestHandler forceBasicAuth({String realm}) {
     if (req.properties.containsKey('user')) return true;
 
     res
-      ..statusCode = HttpStatus.UNAUTHORIZED
-      ..headers[HttpHeaders.WWW_AUTHENTICATE] =
-          'Basic realm="${realm ?? 'angel_auth'}"'
+      ..statusCode = 401
+      ..headers['www-authenticate'] = 'Basic realm="${realm ?? 'angel_auth'}"'
       ..end();
     return false;
   };
 }
 
-/// Restricts access to a resource via authentication.
-Future<bool> requireAuth(RequestContext req, ResponseContext res,
-    {bool throwError: true}) async {
-  bool _reject(ResponseContext res) {
-    if (throwError) {
-      res.statusCode = HttpStatus.FORBIDDEN;
-      throw new AngelHttpException.forbidden();
-    } else
-      return false;
-  }
+/// Use [requireAuthentication] instead.
+@deprecated
+final RequestMiddleware requireAuth = requireAuthentication(userKey: 'user');
 
-  if (req.properties.containsKey('user') || req.method == 'OPTIONS')
-    return true;
-  else
-    return _reject(res);
+/// Restricts access to a resource via authentication.
+RequestMiddleware requireAuthentication({String userKey: 'user'}) {
+  return (RequestContext req, ResponseContext res,
+      {bool throwError: true}) async {
+    bool _reject(ResponseContext res) {
+      if (throwError) {
+        res.statusCode = 403;
+        throw new AngelHttpException.forbidden();
+      } else
+        return false;
+    }
+
+    if (req.properties.containsKey(userKey) || req.method == 'OPTIONS')
+      return true;
+    else
+      return _reject(res);
+  };
 }