+++ lib/src/strategy.dart
@@ -7,7 +7,8 @@ abstract class AuthStrategy {
   String name;
 
   /// Authenticates or rejects an incoming user.
-  Future authenticate(RequestContext req, ResponseContext res, [AngelAuthOptions options]);
+  Future authenticate(RequestContext req, ResponseContext res,
+      [AngelAuthOptions options]);
 
   /// Determines whether a signed-in user can log out or not.
   Future<bool> canLogout(RequestContext req, ResponseContext res);