+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.2.0
+* Deprecate `requireAuth`, in favor of `requireAuthentication`.
+* Allow configuring of the `userKey`.
+* Add `authenticateAndContinue`.
+* Deprecate `middlewareName`.
+
 # 1.1.1+6
 * Fix a small logic bug that prevented `LocalAuthStrategy`
 from correctly propagating the authenticated user when